+++ src/main/resources/db/changelog/00006_order_table_create.sql
@@ -1,13 +1,13 @@
 CREATE TABLE orders (
-    id BIGINT PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     moment TIMESTAMP,
-    customer_id serial,
+    customer_id INT,
     CONSTRAINT fk_customer FOREIGN KEY (customer_id) REFERENCES app_user(id)
 );
 
 
 CREATE TABLE order_items (
-    order_id BIGINT,
+    order_id INT,
     product_id INT,
     quantity INT,
     price REAL,
